V1 of n8 Bit Special
This is the V1 hardware schematic + PCBs that were actually printed. This includes a *hardware defect* where voltage was not stable bc the LED caused a voltage drop. This is fixed in V2 via an additional buffering transitor (hopefully) that could get backported to make V1 functional. The ports are also too close together to really be able to use all 4 ports unless no sockets are used and the chips on the board are very thin. This is also addressed in V2. But I think it important to at least keep the original version.

Lastly, the GAL logic here _does_ include a fix. Previously, both RAM_CS signals were _not qualified_ with phi2. This is bad! So this version includes a GAL glue logic code update to qualify the RAM CS signals with CLK being high. But since this doesn't actually require a hardware revision, I included the fix here. If you want the buggy logic back to get my full experience, you can remove the `CLK &` in the PLD file.